fix(cell): honor dispatch boundary (eng/go §2.18) — extract cellrun; use slices.Contains
CI's dispatch-boundary guard (INVARIANTS.md T-002, eng/go §2.18) failed:
internal/cli/cmd_cell_run.go carried IO/JSON/arg-parsing domain logic and
imported os + encoding/json in the dispatch layer.

- Extract the run logic into a new domain package internal/cellrun (Run over
  argv + stdio, returning an exit code); cmd_cell_run.go is now a thin
  dispatch wrapper delegating to cellrun.Run.
- Move the command-level tests to internal/cellrun/run_test.go.
- Replace the hand-rolled contains() with slices.Contains in cellspec (eng/go
  §2.11 stdlib-first).

Local dispatch-guard grep passes; gofmt/build/vet/go test -race green across
the module; cn cell run smoke unchanged (accepted, exit 0).
